refactor!: The `protocolVersion` in the response is now represented as an `Integer`. (#647)

+++ src/main/java/org/stellar/sdk/responses/RootResponse.java
@@ -40,7 +40,7 @@ public class RootResponse extends Response {
   Integer currentProtocolVersion;
 
   @SerializedName("supported_protocol_version")
-  Long supportedProtocolVersion;
+  Integer supportedProtocolVersion;
 
   @SerializedName("core_supported_protocol_version")
   Integer coreSupportedProtocolVersion;
